+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+- Prop `striped` for add striped rows in the `Table` component.
+
 ## [9.106.2] - 2020-01-27
 
 ### Fixed